style: drop three more redundant casts flagged by clang-tidy
Same readability-redundant-casting class as the previous commit, in
files the offsetof sweep pulled into lint scope:

- Collection.c: index is already zend_ulong
- DefaultIndex.c: value_str_length is already size_t
- Timestamp.c: the outer cast wraps an expression that is already double

Verified by running the lint job's own command over all 38 changed
files locally: zero errors.

+++ src/Database/DefaultIndex.c
@@ -129,7 +129,7 @@ void php_scylladb_index_build_option(php_scylladb_index *index)
       if (cass_value_get_string(value, &value_str, &value_str_length) != CASS_OK) {
         continue;
       }
-      add_assoc_stringl_ex(&index->options, key_str, key_str_length, (char *)(value_str), (size_t)(value_str_length));
+      add_assoc_stringl_ex(&index->options, key_str, key_str_length, (char *)(value_str), value_str_length);
     }
   }
 }
